Remove requirement of providing an explicit staking target (#20)
This PR helps remove the requirement to provide an explicit staking
target (integrator contract address for ethereum partial staking and
validator address for sol) while making staking api calls.

As we move to a self-service world, we want customers to be able to just
come and stake to default staking targets i.e. default integration
contracts on both Holesky and Mainnet in the case of Partial ETH staking
and default coinbase validators on both Devnet and Mainnet in the case
of sol staking.

As a result, we make the integrator contract address validator address
an optional filed. When not set, the backend defaults that network's
default staking target. For Partial ETH staking, customers that have
worked with us to get a dedicated contract deployed, we default to their
specific integrator contract address.