Prep 0.7.1 release (#22)
This is a dummy PR to just bump a patch version to get a clean GH tag,
release and npm pkg published. Previous PR to automate GH release
process mistakenly published the npm pkg. I have since removed the 0.7.0
tags, release and npm version but the mainline build still fails at
re-publishing the 0.7.0 npm pkg. This is to help get a clean start.