Client interface updates (#31)
## Description Of Change

The ts client currently has an interface where for APIs like
`getWorkflow` and `performWorkflowStep` it expects the user to pass a
workflow id. This forces a user to parse out the workflow.name and get
it from there. Instead what's easier is, the client interface takes the
workflow name as input. One less thing the caller needs to worry about.

## Testing Procedure

<!-- Describe how this change has been verified. Either via new
automated tests or manual testing -->

Ran `ts-node examples/solana/create-and-process-workflow.ts` with
appropriate input, to verify if all works.