Add Test Suite, Achieve 80% Coverage (#41)
## Description Of Change

This PR adds a testing suite to this repository, achieving 80%+ test
coverage of all relevant client files.

While working on the test suite, I ran across an old method on the
client `performWorkflowStep` which was broken. As far as we can tell, no
one uses this method and it doens't fit into the API surface of our
Staking API anymore. So I've gone ahead and removed it. This makes this
PR a breaking change.

I attempted to modify as little actual client code as possible while
making these changes, focusing only on the testing and the appropriate
coverage. Slight alterations were made, where necessary, to achieve
maximal coverage (ex: exporting a method).

As you can see in the image attached in a comment, we've achieved 80%+
on statements and lines. We can still improve coverage for branches and
funcs. But considering we started at 0 across the board, this is a huge
improvement. The coverage image I posted can be regenerated via `npm run
coverage`.

## Testing Procedure

Added dozens of tests and achieved 80%+ test coverage :)